Replaced memset/memcpy on struct item with direct assignment
This adds better type checking at compile time. Calling memcpy with
sizeof(struct foo) can lead to subtle errors if the type of one (or
both) of the arguments is changed. The same applies to initializations
using memset and sizeof(struct foo), which could potentially go beyond
the struct's boundaries (or only initialize part of it) if the type of
the argument is changed.

This should not be a performance concern with recent compilers, since
in general, the direct assignment can be better optimized, since the
intent is made clear to the compiler (and it can inline appropriate
memory copy operations for sufficiently small structs), while in the
worst case it calls memcpy behind the scenes.

Signed-off-by: Haru <haru@dotalux.com>